% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_period.R
\name{academic_period}
\alias{academic_period}
\title{Academic period}
\usage{
academic_period(x, type = "period", date_table = NULL)
}
\arguments{
\item{x}{A date, or vector with multiple dates. POSIXct is also accepted.}

\item{type}{Default: "period": Returns the academic period.
at "year" the academic year is returned.}

\item{date_table}{A table in which the correct academic year and period can
be found. When the argument is null, Dates.csv from vvconverter will be used.}
}
\value{
The academic year or period in which the specified date falls.
}
\description{
In this function a date is translated to the academic year or period
in which that date falls. This is based on the periods contained in the file
Dates.csv which is an given dataset in the package vvconverter.
This information is entered manually maintained on the basis of the
academic year calendar of the VU. However, the user can also use his own created
dataset, with the same column names, by using the argument date_table
}
\concept{vector calculations}
