% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quality_report.R
\name{add_attribute}
\alias{add_attribute}
\title{Adjust dataframe for dataMaid
Automatically add based on description and category in documentation
the attribute shortDescription and label to the variables of the Analysis Set}
\usage{
add_attribute(x, varname, Documentation)
}
\arguments{
\item{x}{vector}

\item{varname}{Variable name from analysis set}

\item{Documentation}{Documentation of the analysis set}
}
\value{
Analysis set with labels
}
\description{
Adjust dataframe for dataMaid
Automatically add based on description and category in documentation
the attribute shortDescription and label to the variables of the Analysis Set
}
