% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_script.R
\name{archive_current_script}
\alias{archive_current_script}
\title{Archive Current Script}
\usage{
archive_current_script()
}
\description{
Function to archive the current script open in the source pane in RStudio.
}
