% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_script.R
\name{archive_script}
\alias{archive_script}
\title{Archive scripts to the archive repo}
\usage{
archive_script(File_path, Archive_path = NULL)
}
\arguments{
\item{File_path}{File(s) to archive.}

\item{Archive_path}{Name of the path to your used Archive. Use in format: "../archive_example".
If NULL, system variables will be used.}
}
\description{
Function to archive script to a locally cloned archive repository.
Both character strings as lists of character vectors are accepted as input.
}
\examples{
archive_script("20. Test/Tomer_3170.R")
archive_script(c("20. Test/Tomer_3209.R", "20. Test/Tomer_23432.R", "20. Test/Tomer_3179.R"))
}
