% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{assert_aanwezige_kolommen}
\alias{assert_aanwezige_kolommen}
\title{Assert Aanwezige kolommen}
\usage{
assert_aanwezige_kolommen(Naming, df, Collectie)
}
\arguments{
\item{Naming}{Naming bestand van assertion}

\item{df}{dataframe dat geassert wordt}

\item{Collectie}{Assertions collectie}
}
\description{
Functie die checkt of de kolomnamen uit het documentatiebestand overeenkomen
met de kolomnamen in het te asserten dataframe
}
