% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assert_dataframe_export}
\alias{assert_dataframe_export}
\title{Assert dataframe export}
\usage{
assert_dataframe_export(
  df,
  kolommen_uniek = c("INS_Studentnummer", "INS_Opleidingsnaam_2002", "INS_Studiejaar",
    "INS_Inschrijvingsjaar")
)
}
\arguments{
\item{df}{data die moet worden geassert}

\item{kolommen_uniek}{Kolommen waar op duplicates wordt geassert. Default is
INS_Studentnummer, INS_Opleidingsnaam_2002, INS_Studiejaar, INS_Inschrijvingsjaar.}
}
\description{
Functie die checkt of data meer dan 0 rijen heeft, geen dubbele rijen heeft,
geen kolommen heeft met enkel 0 en geen kolommen heeft met enkel NA.
}
