% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assert_logical_named}
\alias{assert_logical_named}
\title{Assert logical named}
\usage{
assert_logical_named(Kolom, df, prefix_kolom = NULL, ...)
}
\arguments{
\item{Kolom}{een Character vector of string met de kolomnaam die getest moet
worden}

\item{df}{het dataframe dat de kolom bevat}

\item{prefix_kolom}{standaard NULL. Als deze opgegeven wordt, wordt in de
assertion message een tekst voor de variabelenaam geplakt.}

\item{...}{De overige parameters worden meegegeven aan de functie assert_logical}
}
\description{
Een functie om checkmate::assert_logical te gebruiken om de assertion uit te
voeren waarmee gecheckt wordt of de waarde een logical type heeft.
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_pattern_table}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{test_naming}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
