% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{assert_naming}
\alias{assert_naming}
\title{Assert naming}
\usage{
assert_naming(
  df,
  Naming,
  Naam_bestand,
  Maak_nieuwe_kolomwaarden = F,
  Assert_met_veldnaam = F,
  Bericht = NULL,
  fail = "stop"
)
}
\arguments{
\item{df}{Dataframe waar over geassert wordt}

\item{Naming}{Het bijbehorende documentatie bestand met assert informatie voor
het dataframe}

\item{Naam_bestand}{De naam van het databestand wat bekeken wordt. Deze naam moet
uniek zijn in het hele project}

\item{Maak_nieuwe_kolomwaarden}{True als de functie nog een andere functie, build_kolomwaarden,
moet aanroepen om een kolommen bestand te maken waarmee subset/list assertions gedaan
kunnen worden.}

\item{Assert_met_veldnaam}{True als de assertions gedaan worden op de vertaalde namen,
dus Veldnaam ipv Veldnaam_export}

\item{Bericht}{Bericht dat bij de assertion message bijgevoegd wordt.}

\item{fail}{"stop" geeft een error, "warning" geeft een warning als de assertions
niet geslaagd zijn.}
}
\description{
Functie voor assertions voor een dataframe.
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_pattern_table}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{test_naming}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
