% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assert_no_duplicates_in_group}
\alias{assert_no_duplicates_in_group}
\title{Assert geen duplicates in groep}
\usage{
assert_no_duplicates_in_group(
  df,
  group_vars = c("INS_Studentnummer", "INS_Opleidingsnaam_2002", "INS_Inschrijvingsjaar"),
  assertion_fail = "stop"
)
}
\arguments{
\item{df}{Het te controleren dataframe}

\item{group_vars}{De groep-variabelen als een character-vector. De default is
de combinatie van INS_studentnummer, INS_Opleidingsnaam_2002 en INS_Inschrijvingsjaar.}

\item{assertion_fail}{Hoe de functie reageert bij een fail. Dit is of een "warning",
waarbij enkel een warning wordt gegeven over de fail, of een "stop", waarbij de
uitvoering van de functie stopt en de melding weergeeft, of een "AssertCollection",
waarbij het fail-bericht aan een assertion collectie wordt toegevoegd.}
}
\description{
Controleer of er per groep precies 1 rij is. Als er meerdere rijen per groep
gevonden zijn, slaagt de assertion niet.
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
