% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assertion_geen_dubbele_rijen}
\alias{assertion_geen_dubbele_rijen}
\title{assertion_geen_dubbele_rijen}
\usage{
assertion_geen_dubbele_rijen(
  df,
  Collectie,
  kolommen_uniek = c("INS_Studentnummer", "INS_Opleidingsnaam_2002", "INS_Studiejaar",
    "INS_Inschrijvingsjaar")
)
}
\arguments{
\item{df}{Dataframe}

\item{Collectie}{lijst}

\item{kolommen_uniek}{Default: c("INS_Studentnummer", "INS_Opleidingsnaam_2002", "INS_Studiejaar", "INS_Inschrijvingsjaar")}
}
\value{
Collectie
}
\description{
Check of er geen dubbele rijen zijn
}
\examples{
\dontrun{
assertion_geen_dubbele_rijen(df, Collectie)
}
}
