% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assertion_geen_kolommen_met_enkel_0}
\alias{assertion_geen_kolommen_met_enkel_0}
\title{assertion_geen_kolommen_met_enkel_0}
\usage{
assertion_geen_kolommen_met_enkel_0(df, Collectie)
}
\arguments{
\item{df}{Dataframe}

\item{Collectie}{Lijst}
}
\value{
Collectie
}
\description{
Check of er geen kolommen met enkel 0 waarden zijn
}
\examples{
\dontrun{
assertion_geen_kolommen_met_enkel_0(dfInspect, Collectie)
}
}
