% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assertion_geen_kolommen_met_enkel_na}
\alias{assertion_geen_kolommen_met_enkel_na}
\title{assertion_geen_kolommen_met_enkel_na}
\usage{
assertion_geen_kolommen_met_enkel_na(df, Collectie)
}
\arguments{
\item{df}{Dataframe}

\item{Collectie}{Lijst}
}
\value{
Collectie
}
\description{
Check of er geen kolommen met enkel NA values zijn
}
\examples{
\dontrun{assertion_geen_kolommen_met_enkel_na(df, Collectie)}
}
