% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assertion_meer_dan_0_rijen}
\alias{assertion_meer_dan_0_rijen}
\title{assertion_meer_dan_0_rijen}
\usage{
assertion_meer_dan_0_rijen(df, Collectie)
}
\arguments{
\item{df}{Dataframe}

\item{Collectie}{lijst}
}
\value{
Collectie
}
\description{
Check of er meer dan 0 rijen in de dataframe zitten.
}
\examples{
\dontrun{assertion_meer_dan_0_rijen(df, Collectie)}
}
