% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{assertion_message}
\alias{assertion_message}
\title{Assertion message}
\usage{
assertion_message(message, assertion_fail = "stop")
}
\arguments{
\item{message}{The assertion message to return.}

\item{assertion_fail}{The type on which the assertion is returned. The possible
ways are "AssertCollection", where the message belongs to an assertion collection
is added, "warning", giving only a warning and the
function continues, or "stop", where the function stops and returns the message.}
}
\description{
Function that returns the assertion fail on the specified type. This can be with
an AssertCollection, a warning or a stop.
}
