% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{bepaal_subtotalen_met_transitiematrix}
\alias{bepaal_subtotalen_met_transitiematrix}
\title{bepaal_subtotalen_met_transitiematrix}
\usage{
bepaal_subtotalen_met_transitiematrix(df, aantal)
}
\arguments{
\item{df}{Dataframe}

\item{aantal}{Soorten uitstroom}
}
\value{
Transitiematrix in lang formaat
}
\description{
Functie om subtotalen te berekenen met een transitiematrix
}
\examples{
\dontrun{bepaal_subtotalen_met_transitiematrix(df, aantal)}
}
