% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_as.R
\name{build_as}
\alias{build_as}
\title{build_as}
\usage{
build_as(
  Analysis_set,
  AS_archive,
  latest,
  files_list,
  complete_rebuild = FALSE
)
}
\arguments{
\item{Analysis_set}{the Analysis set}

\item{AS_archive}{the archived version of the analysis set}

\item{latest}{latest date analysis set was built}

\item{files_list}{list of files to run}

\item{complete_rebuild}{logical, set to TRUE to always completely rebuild AS, regardless of changes}
}
\value{
Analysis set
}
\description{
A function used to update the analysis set based on an archived version
and a list containing r scripts.
}
