% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{build_kolomwaarden}
\alias{build_kolomwaarden}
\title{Build kolomwaarden}
\usage{
build_kolomwaarden(df, Naming, Naam_bestand)
}
\arguments{
\item{df}{Dataframe waar de kolomnamen van geanalyseerd worden.}

\item{Naming}{Een tabel of dataframe met de veldnamen van kolommen van het
desbetreffende dataframe met meer informatie.}

\item{Naam_bestand}{Naam van het bestand waar de kolomwaarden voor gebouwd worden,
zonder bestandssoort (dus enkel "Namen" of "Gegevens")}
}
\description{
Functie om unieke waarden van kolommen te vinden, en deze, met de bijbehorende
kolomnaam, weg te schrijven naar een csv bestand.
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
