% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{check_assertions}
\alias{check_assertions}
\title{Check assertions}
\usage{
check_assertions(
  df,
  Naming,
  Naam_bestand,
  Kolomwaarden = NULL,
  build_nieuwe_kolommen = F,
  Assert_na_vertaling = F,
  fail = "stop",
  Bericht = NULL,
  Test_documentatie = FALSE,
  Outputmap = NULL
)
}
\arguments{
\item{df}{Dataframe waar over geassert wordt}

\item{Naming}{Het bijbehorende documentatie bestand met assert informatie voor
het dataframe}

\item{Naam_bestand}{De naam van het databestand wat bekeken wordt. Deze naam moet
uniek zijn in het hele project}

\item{Kolomwaarden}{de waarden die mogelijk in een kolom voor kunnen komen}

\item{build_nieuwe_kolommen}{True als de functie nog een andere functie, build_kolomwaarden,
moet aanroepen om een kolommen bestand te maken waarmee list assertions gedaan
kunnen worden.}

\item{Assert_na_vertaling}{True als de assertions gedaan worden op de vertaalde namen,
dus Veldnaam ipv Veldnaam_export}

\item{fail}{"stop" geeft een error, "warning" geeft een warning als de assertions
niet geslaagd zijn.}

\item{Bericht}{Bericht dat bij de assertion message bijgevoegd wordt.}

\item{Test_documentatie}{Boolean of testdocumentatie gebruikt moet worden of
assertdocumentatie}

\item{Outputmap}{De Outputmap van het bestand, bijvoorbeeld 1. Ingelezen data}
}
\description{
Functie voor assertions voor een dataframe.
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
