% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_vvsculptor.R
\name{check_common_cols}
\alias{check_common_cols}
\title{check for common columns between dataframes.}
\usage{
check_common_cols(...)
}
\arguments{
\item{...}{dataframes}
}
\value{
A character vector containing column names.
}
\description{
This function checks whether multiple dataframes have overlapping column names.
}
