% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_education_name.R
\name{check_education_name}
\alias{check_education_name}
\title{check education name}
\usage{
check_education_name(df, actual_edu_names = NULL, col_edu_names = NULL)
}
\arguments{
\item{df}{A data frame}

\item{actual_edu_names}{a column with the actual education names, default is null,
than it will check for system variables}

\item{col_edu_names}{a column with the education names that you want to check,
default is null, than it will check for system variables}
}
\description{
Check whether the specified dataframe uses education names that
do not appear in the original education names table.
If there are other training name names, an error is returned.
If the training names all match, a confirmation is given.
}
\seealso{
Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{tests}
