% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_name.R
\name{check_name}
\alias{check_name}
\title{check name}
\usage{
check_name(string)
}
\arguments{
\item{string}{The variable name. Can also be a list of names.}
}
\value{
The result of the check. This can be: "Correct","spaces in name",
"Wrong characters in name","No 3-letter prefix",
"Word after prefix has no capital letter"
and "Words after first word have capital letters"
}
\description{
This function checks if a variable name is correct following the style guide.
}
