% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_filenames.R
\name{check_naming_convention_from_yaml}
\alias{check_naming_convention_from_yaml}
\title{Check naming conventions from a YAML file}
\usage{
check_naming_convention_from_yaml(x, yaml_file)
}
\arguments{
\item{x}{A character string representing the file name to check.}

\item{yaml_file}{A character string representing the path to the YAML file containing the naming conventions.}
}
\value{
A list of logical values indicating whether each naming convention is met by the file name.
}
\description{
This function reads a YAML file containing naming conventions and checks if a given file name adheres to these conventions.
}
