% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unused_objects.R
\name{check_unused_objects}
\alias{check_unused_objects}
\title{Identify unused objects in a script file}
\usage{
check_unused_objects(filepath)
}
\arguments{
\item{filepath}{A character string specifying the path to the script file.}
}
\value{
A character vector of unused objects in the script, or NA if no unused objects are found.
}
\description{
This function reads a script file, identifies objects that are assigned but not used,
and returns a list of unused objects. If no unused objects are found, it returns NA.
}
