% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_objects.R
\name{clear_global}
\alias{clear_global}
\title{Clear global.}
\usage{
clear_global(...)
}
\arguments{
\item{...}{which object(s) to keep.}
}
\description{
Clear global environment except for project specific objects.
}
