% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_objects.R
\name{clear_global_proj}
\alias{clear_global_proj}
\title{Clears the global environment except for objects specified in the DEFAULT_KEEP_LIST system variable.}
\usage{
clear_global_proj()
}
\description{
This function removes all objects from the global environment except for those specified
in the DEFAULT_KEEP_LIST system variable. The variable is expected to be a space-separated
list of object names. If the variable is not set, the function will not run.
}
\note{
This function should be used with caution as it modifies the global environment.
}
\seealso{
\code{\link[base]{rm}} for the underlying mechanism used to remove objects.
}
\keyword{internal}
