% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_objects.R
\name{clear_script_objects}
\alias{clear_script_objects}
\title{clear script objects.}
\usage{
clear_script_objects(
  ...,
  filepath = NULL,
  list = character(),
  pos = -1,
  envir = as.environment(pos),
  line_start = 0,
  line_end = -1L,
  silent = TRUE
)
}
\arguments{
\item{...}{which object(s) to keep.}

\item{filepath}{path to script}

\item{list}{list of objects}

\item{pos}{what position in the environment to clear}

\item{envir}{which environment to clear}

\item{line_start}{from which line}

\item{line_end}{until which line}

\item{silent}{whether to mute console output}
}
\description{
Clear objects which are created in the current script.
}
