% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{controleer_na_waarden}
\alias{controleer_na_waarden}
\title{Controleer NA waarden}
\usage{
controleer_na_waarden(Kolomwaarde, df, Naming, add, fail = "stop")
}
\arguments{
\item{Kolomwaarde}{De kolom uit het dataframe waar voor op NAs moet worden gecheckt}

\item{df}{Dataframe waaron moet worden gezocht}

\item{Naming}{Documentatiebestand waarin de gegevens staan}

\item{add}{Assertion collectie waar assertion bericht aan moet worden toegevoegd}

\item{fail}{Vermeld of asserion bij een failure moet stoppen ("stop") of waarschuwen ("warning")}
}
\description{
Deze functie kijkt voor een kolom of het aantal NAs binnen de grenswaarde ligt
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
