% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_names.R
\name{correct_file_name}
\alias{correct_file_name}
\title{Correct file name}
\usage{
correct_file_name(file_name)
}
\arguments{
\item{file_name}{the File name}
}
\value{
The corrected file name
}
\description{
This function changes the name of the filename variable
(compared to correct_var_name() the underscore is preserved).
}
