% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_names.R
\name{correct_var_name}
\alias{correct_var_name}
\title{Correct var name}
\usage{
correct_var_name(var_name)
}
\arguments{
\item{var_name}{The variable name}
}
\value{
The corrected variable name
}
\description{
This function changes the name of the variable for plots so that prefixes are removed
become.
}
