% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rows_analysis_set.R
\name{count_rows_analysis_set}
\alias{count_rows_analysis_set}
\title{count rows Analysis set}
\usage{
count_rows_analysis_set(
  Dataset,
  ScriptName,
  Number_lines_begin,
  type = "Analysisset",
  slack = T
)
}
\arguments{
\item{Dataset}{The dataset to be tested}

\item{ScriptName}{Name The script name in which this function is executed. This one
filename will be placed in message if number of rows is not equal}

\item{Number_lines_begin}{The number of rows to calculate any difference}

\item{type}{The name of the data file}

\item{slack}{If TRUE(default), the message is also sent to Slack.}
}
\description{
Function to check if the number of rows of a data frame matches
the specified number of rows.
}
