% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_outflow_study_year.R
\name{determine_outflow_study_year}
\alias{determine_outflow_study_year}
\title{determine_outflow_study_year}
\usage{
determine_outflow_study_year(
  SUC_Type_uitstroom,
  INS_Studiejaar,
  OPL_Studielast_nominaal
)
}
\arguments{
\item{SUC_Type_uitstroom}{Soort uitstroom}

\item{INS_Studiejaar}{Jaar van uitstroom}

\item{OPL_Studielast_nominaal}{Nominale hoeveelheid studielast (EC)}
}
\description{
Bepaal uitstroom voor het betreffende studiejaar obv
uiteindelijke uitstroom en studiejaar
}
