% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_inspect.R
\name{df_inspect}
\alias{df_inspect}
\title{Inspect given data frame in Student Analytics VU.}
\usage{
df_inspect(data, print_results, dump = F, dump_location = NULL)
}
\arguments{
\item{data}{data frame}

\item{print_results}{if FALSE then there is not a print in the console, TRUE by default.}

\item{dump}{Dump the output to the XX. Inspect folder}

\item{dump_location}{Path with location to save the dump file, default is NULL}
}
\value{
Metrics data frame
}
\description{
For each variable it returns: Quantity and percentage of zeros (q_zeros and p_zeros respectevly). Same metrics for NA values (q_NA/p_na), and infinite values (q_inf/p_inf). Last two columns indicates data type and quantity of unique values.
The standard deviation, variance, Q1/Median/Q3 & IQR, Min/Max and Mean.
This function prints and returns the results. It is based on the funModeling package.
}
