% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_add}
\alias{documentation_add}
\title{documentation add}
\usage{
documentation_add(data, overwrite = F)
}
\arguments{
\item{data}{Een data frame met de nieuwe rijen voor in het documentatiebestand}

\item{overwrite}{FALSE: de rijen worden toegevoegd, zonder dat de dubbele rijen eerst verwijderd worden.
Als TRUE overschrijven de nieuwe rijen de oorspronkelijke rijen.}
}
\description{
Functie om rijen toe te voegen aan het documentatiebestand. Vervolgens wordt deze geupdate versie
opgeslagen.
}
