% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_add_csv}
\alias{documentation_add_csv}
\title{documentation add csv}
\usage{
documentation_add_csv(dataloc = "/documentation/Update.csv", overwrite = F)
}
\arguments{
\item{dataloc}{De locatie van een databestand met de nieuwe rijen voor in het documentatiebestand}

\item{overwrite}{FALSE: de rijen worden toegevoegd, zonder dat de dubbele rijen eerst verwijderd worden.
Als TRUE overschrijven de nieuwe rijen de oorspronkelijke rijen.}
}
\description{
Wrapper rond documentation_add(), hier kan een locatie van een csv opgegeven worden. Deze wordt
toegevoegd aan het documentatiebestand.
}
