% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_change}
\alias{documentation_change}
\title{documentation change}
\usage{
documentation_change(old, new)
}
\arguments{
\item{old}{de oude naam die veranderd moet worden}

\item{new}{de nieuwe naam die veranderd moet worden, moet tussen "" ingevuld worden
te verwijderen.}
}
\description{
Functie om rijen te verwijderen uit het documentatiebestand. Vervolgens wordt deze geupdate versie
opgeslagen. De verwijderde rijen worden opgeslagen in /documentation/XX. Archief/
}
