% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_dump_csv}
\alias{documentation_dump_csv}
\title{documentation dump csv}
\usage{
documentation_dump_csv(search, dataloc = "/documentation/", ...)
}
\arguments{
\item{search}{De zoekterm om te zoeken in het documentatiebestand.}

\item{dataloc}{De netwerklocatie waar de dump geplaatst wordt.}

\item{...}{Deze parameters worden doorgegeven aan documentation_search()}
}
\description{
Wrapper rond documentation_search(). Het resultaat van de zoekterm wordt opgeslagen in een csv bestand.
}
