% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_generate.R
\name{documentation_generate}
\alias{documentation_generate}
\title{Documentation generate}
\usage{
documentation_generate(
  data,
  description,
  source,
  privacy = "no",
  extra_info = NA
)
}
\arguments{
\item{data}{A data frame containing the variable to be documented}

\item{description}{The description of the variable}

\item{source}{The source of the data}

\item{privacy}{The privacy sensitivity of this variable. Possibilities are:
"no" (default), "yes", "Yes, disability" and "Yes, ethnicity".}

\item{extra_info}{Optionally, extra information can be added}
}
\description{
A function to create a record in the data frame based on a variable in a data frame
generate documentation file. Category and common values become
completed based on the prefix of the variable name and the occurring values
in the dates.
}
