% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation_get.R
\name{documentation_get}
\alias{documentation_get}
\title{Get documentation analysis set}
\usage{
documentation_get(AS_doc = NULL)
}
\arguments{
\item{AS_doc}{Path to the documentation file of the analysisset}
}
\description{
Function to load the latest version of the documentation.
}
