% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_remove}
\alias{documentation_remove}
\title{documentation remove}
\usage{
documentation_remove(remove)
}
\arguments{
\item{remove}{een string, of vector met strings om de variabelen uit de documentatie
te verwijderen.}
}
\description{
Functie om rijen te verwijderen uit het documentatiebestand. Vervolgens wordt deze geupdate versie
opgeslagen. De verwijderde rijen worden opgeslagen in /documentation/XX. Archief/
}
