% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_save}
\alias{documentation_save}
\title{documentation save}
\usage{
documentation_save(data)
}
\arguments{
\item{data}{Het documentatiebestand in de vorm van een data frame}
}
\description{
Functie om een nieuw documentatiebestand op te slaan in /documentation/ en in het archief. Opslaan is alleen
mogelijk als er geen duplicaten in de kolom "Veldnaam" zitten.
}
