% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_search}
\alias{documentation_search}
\title{documentation search}
\usage{
documentation_search(
  search,
  output = "veldnaam",
  omschrijving = T,
  veldnaam = T,
  categorie = T
)
}
\arguments{
\item{search}{Zoekterm waarop de documentatie doorzocht wordt}

\item{output}{"veldnaam", voor een vector met alle gevonden veldnamen. "tabel" voor
een tabel met alleen Categorie, Veldnaam, en Omschrijving. "all" voor alle informatie}

\item{omschrijving}{T/F indien hierin wel of niet in gezocht moet worden}

\item{veldnaam}{T/F indien hierin wel of niet in gezocht moet worden}

\item{categorie}{T/F indien hierin wel of niet in gezocht moet worden}
}
\description{
Doorzoekt de documentatie aan de hand van de opgegeven zoekterm. Als
resultaat wordt de documentatie van de gevonden kolommen teruggegeven.
}
