% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation_search_output}
\alias{documentation_search_output}
\title{documentation search output}
\usage{
documentation_search_output(Zoekresultaat, output, search)
}
\arguments{
\item{Zoekresultaat}{Een data frame met de variabelen die als output geleverd moeten
worden}

\item{output}{"veldnaam", voor een vector met alle gevonden veldnamen. "tabel" voor
een tabel met alleen Categorie, Veldnaam, en Omschrijving. "all" voor alle informatie}

\item{search}{De zoekterm}
}
\description{
Om de output van het documentatiebestand te message.
Deze functie wordt in principe alleen opgeroepen vanuit de functie documentation_search()
}
