% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{download_github_file}
\alias{download_github_file}
\title{Download a file from a GitHub repository}
\usage{
download_github_file(url, temp_file_path)
}
\arguments{
\item{url}{A character string specifying the URL of the file to download.}

\item{temp_file_path}{A character string specifying the path to the temporary
file where the downloaded content will be saved.}
}
\value{
A character string indicating the path to the temporary file where the
downloaded content has been saved.
}
\description{
This function takes a URL pointing to a raw file in a GitHub repository and a
temporary file path where the downloaded content will be saved. It makes a GET
request to the provided URL and, if successful, writes the raw content of the
response to the specified temporary file path.
}
