% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{dump_subset_to_file}
\alias{dump_subset_to_file}
\title{dump subset to file}
\usage{
dump_subset_to_file(
  df,
  studentnummer,
  dataloc = "Exports/Student_Exports/",
  suffix = NULL
)
}
\arguments{
\item{df}{Het databestand met de analyseset}

\item{studentnummer}{Het studennummer als integer}

\item{dataloc}{De locatie waar het bestand naar weggeschreven wordt}

\item{suffix}{Een suffix dat achter de bestandsnaam geplaatst wordt}
}
\value{
een xlsx-bestand met een subset obv 1 studentnummer
}
\description{
Data uit de analyseset wordt van de opgegeven student geexporteerd naar een csv
NA's worden verwijderd.
}
