% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_subset_to_testfile.R
\name{dump_subset_to_testfile}
\alias{dump_subset_to_testfile}
\title{Dump subset to testfile}
\usage{
dump_subset_to_testfile(
  df,
  studentnummer = NULL,
  output = "XX. Test/",
  dataloc,
  suffix = NULL
)
}
\arguments{
\item{df}{The data file containing the analysis set}

\item{studentnummer}{number The student number as integer. A single value or a vector with multiple
values.}

\item{output}{The location in the output directory where the file will be written.}

\item{dataloc}{optional: If no output directory is specified, a
data location on the network drive can be specified}

\item{suffix}{A suffix that is placed after the file name}
}
\value{
A csv file with a subset based on 1 or more student numbers
}
\description{
Data from the analysis set is exported from the specified student to a
csv. The file name contains a suffix with the date and time.
}
