% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{dump_to_testfile}
\alias{dump_to_testfile}
\title{dump to testfile}
\usage{
dump_to_testfile(df, output = "XX. Test/", dataloc, suffix = NULL)
}
\arguments{
\item{df}{Het databestand met de analyseset}

\item{output}{De locatie in de output-directory waar het bestand naar weggeschreven wordt.}

\item{dataloc}{optioneel: Als er geen output directory is opgegeven, kan hier een
datalocatie op de netwerkschijf opgegeven worden}

\item{suffix}{Een suffix dat achter de bestandsnaam geplaatst wordt}
}
\value{
Het databestand wordt weggeschreven als csv-bestand
}
\description{
Data wordt geexporteerd naar een csv. In de bestandsnaam staat een suffix
met de datum en tijd.
}
