% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_results}
\alias{export_results}
\title{Export Results.R}
\usage{
export_results(Analysis_file, Results_location = NULL, Analysis_set = NULL)
}
\arguments{
\item{Analysis_file}{This must be a data file of a subset of the analysis set.
The script provides the results of the student numbers that are present
are in this file.}

\item{Results_location}{The path to your registered results. If NULL, system
variables are used}

\item{Analysis_set}{An Analysis_set that is in the format of the VUSA analysis set,
wit the columns: INS_Opleidingsnaam_2002", "INS_Faculteit", "INS_Studentnummer",
"INS_Inschrijvingsjaar", and "RES_Academisch_jaar_beoordeling". If NULL,
system variables are used}
}
\description{
This function creates a data frame with the results for each student
in the entered analysis set. NB! To compare it uses an standart VUSA
Analysis set, together with its columns: INS_Opleidingsnaam_2002",
"INS_Faculteit", "INS_Studentnummer", "INS_Inschrijvingsjaar", and
"RES_Academisch_jaar_beoordeling". So make sure you have this columns present
in your Analysis_set when using this function
}
\details{
Dependencies: Index.R
Export Analysis Set.R
}
