% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{extract_filepaths_readrds_csv}
\alias{extract_filepaths_readrds_csv}
\title{Extraheer bestandspaden ReadRDS CSV}
\usage{
extract_filepaths_readrds_csv(bestandspad)
}
\arguments{
\item{bestandspad}{Het complete bestandspad van het script}
}
\description{
Functie die voor Genereer_outputbestanden_database bestandspaden achterhaald van ingelezen bestanden
}
