% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_save.R
\name{filter_settings_on_type}
\alias{filter_settings_on_type}
\title{Filter Settings Based on Type}
\usage{
filter_settings_on_type(settings_script_dir, settings_type)
}
\arguments{
\item{settings_script_dir}{Data frame containing the settings script directory.}

\item{settings_type}{Character, the type of setting to filter on. If NULL, the function will select the first type.}
}
\value{
A data frame containing the filtered settings script directory.
}
\description{
Filters the settings script directory based on the supplied settings type.
Handles edge cases and provides appropriate messages and errors.
}
