% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pattern_r.R
\name{find_pattern_csv}
\alias{find_pattern_csv}
\title{Find pattern in csv files.}
\usage{
find_pattern_csv(
  pattern = "Hello World",
  path = ".",
  case.sensitive = TRUE,
  show.results = TRUE
)
}
\arguments{
\item{pattern}{Pattern to search on}

\item{path}{de folder waar gezocht moet worden, default working directory}

\item{case.sensitive}{Whether pattern is case sensitve or ntot}

\item{show.results}{whether to show the results in the console}
}
\value{
dataframe met R script paths
}
\description{
Function to search for a pattern in csv files.
}
\examples{
\dontrun{
find_pattern_csv(pattern = "croho")
find_pattern_csv("croho", case.sensitive = FALSE)
}
}
