% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_vvauditor.R
\name{find_sys_variables}
\alias{find_sys_variables}
\title{Find System Variables in a Package}
\usage{
find_sys_variables(package_name)
}
\arguments{
\item{package_name}{The name of the R package to scan for system variable usage.}
}
\value{
A data frame with two columns: \code{Function} and \code{System_Variable}. Each row represents a function within the package that uses a system variable, along with the name of the variable.
}
\description{
This function scans a specified R package for calls to \code{Sys.getenv()} and extracts the names of system variables used within the package's functions.
It returns a data frame listing the functions that use system variables and the names of those variables.
}
