% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vvmover_create_documentatie.R
\name{first_class}
\alias{first_class}
\title{First class}
\usage{
first_class(x)
}
\arguments{
\item{x}{Object waar de class van wordt gezocht}
}
\description{
Functie die de class van een element of vector zoekt en hier, indien er meerdere
benamingen van het class zijn, de eerste class van neemt.
}
