% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysisset.R
\name{get_analysisset}
\alias{get_analysisset}
\title{Get the analysis set from the network drive}
\usage{
get_analysisset(columns = NULL, AS_path = NULL)
}
\arguments{
\item{columns}{List of specific columns to read in.}

\item{AS_path}{Path of the location for the analysis set}
}
\value{
Analysis set 1, 1R = random set, 7 or 7 Extra features from the branch set
}
\description{
Get the analysis set from the network drive
}
