% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysisset_7.R
\name{get_analysisset_7}
\alias{get_analysisset_7}
\title{Get Analysis set 7}
\usage{
get_analysisset_7(
  cols = NULL,
  cols_match = c("INS_Studentnummer", "INS_Opleidingsnaam_2002", "INS_Inschrijvingsjaar"),
  AS_7 = NULL
)
}
\arguments{
\item{cols}{Variable name as text value, or vector with multiple
text values of variables to be added to Analysis Set 7}

\item{cols_match}{Matching columns for joining AS_1 and AS_7, the default
columns are: "INS_Studentnummer","INS_Opleidingsnaam_2002","INS_Inschrijvingsjaar"}

\item{AS_7}{Path for the location of Analysis set 7}
}
\description{
Analysis set 7 is retrieved, and the columns specified are from
Analysis set 1 added to Analysis set 7.
}
