% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysisset_vak.R
\name{get_analysisset_vak}
\alias{get_analysisset_vak}
\title{Get the analysis set from the network drive}
\usage{
get_analysisset_vak(columns = NULL, AS_path = NULL)
}
\arguments{
\item{columns}{List of specific columns to read in.}

\item{AS_path}{Path of the location for the analysis set}
}
\value{
Vak analysis set
}
\description{
Get the analysis set from the network drive
}
