% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archived_result_files.R
\name{get_archived_result_files}
\alias{get_archived_result_files}
\title{Get the archived result files to calculate the end results}
\usage{
get_archived_result_files(academic_year, last_date = 0)
}
\arguments{
\item{academic_year}{The academic year that we are looking at}

\item{last_date}{The date of the last checked result file (Date object)}
}
\description{
Get the archived result files to calculate the end results
}
