% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_proj.R
\name{get_last_n_commit_info}
\alias{get_last_n_commit_info}
\title{Retrieve the last N commit information for a file}
\usage{
get_last_n_commit_info(filepath, n = 1)
}
\arguments{
\item{filepath}{A character string specifying the path to the file for which
commit information is to be retrieved.}

\item{n}{An integer specifying the number of commit information to retrieve.
The default is  1, which retrieves the most recent commit information.}
}
\value{
A flat list of commit information, including the commit hash, author,
and message for the last N commits.
}
\description{
This function uses the \verb{git log} command to retrieve the last N commit
information for a specified file. The information includes the commit hash,
author, and commit message.
}
\note{
This function currently only works on Windows systems due to the use of
the \code{shell()} function.
}
\examples{
\dontrun{
# Retrieve the last commit information for a file
get_last_n_commit_info(filepath = "/path/to/file.R")

# Retrieve the last  5 commit information for a file
get_last_n_commit_info(filepath = "/path/to/file.R", n =  5)
}

}
