% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vvmover_create_documentatie.R
\name{get_ratio}
\alias{get_ratio}
\title{get_ratio}
\usage{
get_ratio(data_vector)
}
\arguments{
\item{data_vector}{column to check}
}
\description{
Get percentages for factor columns
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
