% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{get_snippets_file}
\alias{get_snippets_file}
\title{Get the path to the snippets file for a given type}
\usage{
get_snippets_file(
  type = c("r", "markdown", "c_cpp", "css", "html", "java", "javascript", "python",
    "sql", "stan", "tex")
)
}
\arguments{
\item{type}{A character string specifying the type of the snippets file.
The default is "r". Other valid types include "markdown",
"c_cpp", "css", "html", "java", "javascript", "python",
"sql", "stan", "tex".}
}
\value{
A character string indicating the path to the snippets file.
}
\description{
This function determines the appropriate path to the snippets file based on
the type of the file and the version of RStudio being used. It handles the
differences in file locations between old and new versions of RStudio.
}
