% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash_var.R
\name{hash_var}
\alias{hash_var}
\title{hash var}
\usage{
hash_var(Vector_to_hash, seed = NULL)
}
\arguments{
\item{Vector_to_hash}{The vector to be hashed.}

\item{seed}{A given seed to hash the vectors, default is NULL}
}
\value{
return_var (=hashed vector)
}
\description{
Hashes a vector.
}
