% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_filepath_code.R
\name{insert_filepath_code}
\alias{insert_filepath_code}
\title{Insert File Path Code}
\usage{
insert_filepath_code()
}
\description{
This function prompts the user to select a file and inserts R code for reading the file based on the selected path.
The code is inserted into the RStudio editor.
}
\examples{
\dontrun{
insert_filepath_code()
}

}
