% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vvmover_create_documentatie.R
\name{is_unieke_identifier}
\alias{is_unieke_identifier}
\title{Is unieke identifier}
\usage{
is_unieke_identifier(Veldnaam, df)
}
\arguments{
\item{Veldnaam}{De naam die getest wordt op uniek zijn}

\item{df}{dataframe waar unieke identifier voor wordt gezocht}
}
\description{
Functie die voor ieder veld kijkt of dit een unieke identifier is.
}
