% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_Tkoppel.R
\name{join_tkoppel}
\alias{join_tkoppel}
\title{Join Tkoppel}
\usage{
join_tkoppel(df, faculteit = F)
}
\arguments{
\item{df}{A data frame containing the numeric column INS_Programme_code_actual,
and to which the training name 2002 must be added.
The program name 2002 will be filled in this position by program names from the program name Z08
column, because these names are the most recent. However, the column will be added under Education Name 2002
because this is the most commonly used column for education names.}

\item{faculteit}{If FALSE (default) the column INS_Faculty is not added.
If TRUE, this column will be added.}
}
\description{
A function to extract the program name (and possibly the faculty) from T Koppel_Z08
to the data frame.
To perform this function it is necessary that in the output folder:
file states: "1. Read data/INS_T couple_Z08.rds"
}
