% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tkoppel_isat.R
\name{join_tkoppel_isat}
\alias{join_tkoppel_isat}
\title{Join Tkoppel isat}
\usage{
join_tkoppel_isat(df, faculteit = F)
}
\arguments{
\item{df}{Een data frame dat de numerieke column INS_Opleidingscode_actueel bevat,
en waaraan de opleidingsnaam 2002 moet worden toegevoegd.}

\item{faculteit}{Als FALSE (default) wordt de column INS_Faculteit niet toegevoegd.
Als TRUE wordt deze column wel toegevoegd.}
}
\value{
df
}
\description{
Een functie om opleidingsnaam (en eventueel de faculteit) uit Tkoppel_isat
toe te voegen aan het dataframe.
Om deze functie uit te voeren is het nodig dat in de outputfolder het volgende
bestand staat: "1. Ingelezen data/INS_Tkoppel_isat.rds"
}
