% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_save.R
\name{load_and_save_config_proj}
\alias{load_and_save_config_proj}
\title{Load and Save Configuration Project}
\usage{
load_and_save_config_proj(find_loc_in_env = TRUE)
}
\arguments{
\item{find_loc_in_env}{if true (the default), a system env variable can be given as a location
within the settings config file can be.}
}
\value{
A data frame of the load and save configurations if the configuration file is valid.
If there are issues with the configuration file, the function issues a warning and returns NULL.
}
\description{
This function is responsible for loading and saving configurations for a project. It retrieves
the configuration path from the system environment variable "load_and_save_config".
}
\note{
For more information on how to use this function, see the vignette:
vignette("Using the load_and_save_config_proj function") and more elaborate documentation at:
https://r.mtdv.me/load_save_config.
}
\examples{
\dontrun{
load_and_save_config_proj()


Create a data frame with the given column names
colnames_config <- c(
  "script_dir",
  "type",
  "load_data_dir",
  "load_base_data_dir",
  "save_data_dir",
  "save_base_data_dir",
  "add_branch",
  "save_csv",
  "save_rds",
  "save_fst",
  "read_extension",
  "load_message",
  "save_message",
  "notes"
)
# Initialize a data frame
load_and_save_config_df <- data.frame(matrix(ncol = length(colnames_config), nrow = 0))
names(load_and_save_config_df) <- colnames_config
# Populate the data frame with some dummy data (you would replace this with your own data)
load_and_save_config_df[1,] <- c(
 "dir/of/script", # Used for str_detect the current path of full script
  NA_character_,
  "specific_load_dir_based_on_script_dir",
  "load/base/dir",
  "specific_save_dir_based_on_script_dir",
  "save/base/dir",
  FALSE,
  TRUE,
  TRUE,
  TRUE,
  "rds",
  NA_character_, # This is shown once per session if filled
  NA_character_, # This is shown once per session if filled
  "Some notes"
)
}
}
