% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doorstroom.R
\name{maak_doorstroomvar_premaster_master}
\alias{maak_doorstroomvar_premaster_master}
\title{maak_doorstroomvar_premaster_master}
\usage{
maak_doorstroomvar_premaster_master(df)
}
\arguments{
\item{df}{Een dataframe}
}
\value{
De boolean INS_Doorstroom_van_premaster_naar_master
}
\description{
Bepaal of de student is doorgestroomd van premaster naar master.
}
\seealso{
Other doorstroom: 
\code{\link{bepaal_doorstroom}()},
\code{\link{pas_doorstroomvariabele_aan}()}
}
\concept{doorstroom}
