% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{maak_inschrijvingen_grid}
\alias{maak_inschrijvingen_grid}
\title{maak_inschrijvingen_grid}
\usage{
maak_inschrijvingen_grid()
}
\value{
Een inschrijvingen grid
}
\description{
Functie om een Inschrijvingen_grid te maken per opleidingsniveau ("B", "P", "M"), faculteit ("BETA", "FGB", "FGW", "FRT", "FSW", "GNK", "RCH", "SBE", "THK"), nominale studielasten (30, 60, 120, 180) en inschrijfmomenten ("Hoofdinschrijving", "Inschrijving na 1 oktober")
}
\examples{
maak_inschrijvingen_grid()
}
