% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{maak_instroomscenario_met_subtotalen_tableau}
\alias{maak_instroomscenario_met_subtotalen_tableau}
\title{maak_instroomscenario_met_subtotalen_tableau}
\usage{
maak_instroomscenario_met_subtotalen_tableau(df, scenario, subscenario = NA)
}
\arguments{
\item{df}{Dataframe}

\item{scenario}{PRG_scenario}

\item{subscenario}{Default: NA}
}
\value{
Matrix instroomscenario met subtotalen klaar voor Tableau
}
\description{
Functie om de matrix om te zetten naar het instroomscenarioformaat voor Tableau
}
\examples{
\dontrun{
maak_instroomscenario_met_subtotalen_tableau(df, scenario, subscenario = NA)
}
}
