% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{maak_transitiematrix_grid}
\alias{maak_transitiematrix_grid}
\title{maak_transitiematrix_grid}
\usage{
maak_transitiematrix_grid()
}
\value{
Transitiematrix van alle bachelor studenten aan de VU
}
\description{
Functie om een Transitiematrix_grid te maken van bachelor studenten aan de VU van de faculteiten "BETA", "FGB", "FGW", "FRT", "FSW", "GNK", "RCH", "SBE", "THK". Er wordt verschil gemaakt tussen "Hoofdinschrijving" en "Inschrijving na 1 oktober" en soorten uitstroom.
}
\examples{
maak_transitiematrix_grid()
}
