% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{maak_transitiematrix_uitstroom}
\alias{maak_transitiematrix_uitstroom}
\title{maak_transitiematrix_uitstroom}
\usage{
maak_transitiematrix_uitstroom(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
dfTransitiematrix_uitstroom
}
\description{
Functie om een Transitiematrix voor uitstroom te maken
}
\examples{
\dontrun{maak_transitiematrix_uitstroom(df)}
}
