% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{maak_transitiematrix_uitstroom_met_subtotalen}
\alias{maak_transitiematrix_uitstroom_met_subtotalen}
\title{maak_transitiematrix_uitstroom_met_subtotalen}
\usage{
maak_transitiematrix_uitstroom_met_subtotalen(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Transitiematrix uitstroom met subtotalen
}
\description{
Functie om de transitieaantallen voor gegroepeerde (zoals Studiejaar, Uitstroom etc.) en geneste data te berekenen
}
\examples{
\dontrun{maak_transitiematrix_uitstroom_met_subtotalen(df)}
}
