% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{maak_transitiematrix_voor_prognosejaren}
\alias{maak_transitiematrix_voor_prognosejaren}
\title{maak_transitiematrix_voor_prognosejaren}
\usage{
maak_transitiematrix_voor_prognosejaren(
  df,
  huidig_academisch_jaar = Huidig_academisch_jaar
)
}
\arguments{
\item{df}{Dataframe}

\item{huidig_academisch_jaar}{Default: 'Huidig_academisch_jaar'}
}
\value{
Transitiematrix voor prognose jaren
}
\description{
Functie om de transitiematrix toe passen voor alle prognosejaren
}
\examples{
\dontrun{
maak_transitiematrix_voor_prognosejaren(df)
}
}
