% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_vvconverter.R
\name{mapping_category}
\alias{mapping_category}
\title{mapping Category}
\usage{
mapping_category(
  Data,
  current,
  new,
  mapping_table_input = NULL,
  mapping_table_name = NULL
)
}
\arguments{
\item{Data}{The data frame.}

\item{current}{Current name.}

\item{new}{New name.}

\item{mapping_table_input}{Defaults to NULL. If a mapping
table must be read.}

\item{mapping_table_name}{Default NULL. If there is an existing mapping table
must be read from the name in the folder}
}
\value{
A new data frame with new values based on the 'mapping tables'.
}
\description{
Analogous to 'mapping_translate', takes a csv from the documentation folder
'mapping tables'. The input of this function is a continuous value, which is mapped
is moved to a category based on intervals.
}
\seealso{
Other mapping: 
\code{\link{mapping_fix}()},
\code{\link{mapping_translate}()}
}
\concept{mapping}
