% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_education_name.R
\name{md_list}
\alias{md_list}
\title{MD list}
\usage{
md_list(x, header = NULL, ordered = F)
}
\arguments{
\item{x}{A vector with a length greater than 0.}

\item{header}{Indicates whether a header is present.}

\item{ordered}{Defaults to FALSE, if TRUE then prefix is "." otherwise "-".}
}
\description{
message a list in markdown.
}
