% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_documentation.R
\name{overwrite_dot_arguments}
\alias{overwrite_dot_arguments}
\title{Overwrite the arguments with the specified dots (...)}
\usage{
overwrite_dot_arguments(function_args, ...)
}
\arguments{
\item{function_args}{The function arguments in a named list}

\item{...}{the ... arguments passed in the function}
}
\value{
a named list of function arguments
}
\description{
This function can be used within a function to ensure that
the ... arguments override the specified function arguments. Thereafter
function arguments can be used with do.call()
}
