% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doorstroom.R
\name{pas_doorstroomvariabele_aan}
\alias{pas_doorstroomvariabele_aan}
\title{Aanpassen doorstroomvariabele o.b.v. bepaal_doorstroom}
\usage{
pas_doorstroomvariabele_aan(df1, df2, bachelor_naar_master = TRUE)
}
\arguments{
\item{df1}{Dataframe gecreeerd met behulp van de functie bepaal_doorstroom(). Dit dataframe bevat de variabelen
INS_Studentnummer, INS_Opleidingsnaam_2002_B, INS_Eerste_jaar_opleiding_en_instelling_B, INS_Doorstroom_van_bachelor_naar_master_B,
of deze variabelen voor de premaster, in dat geval is de suffix _P in plaats van _B.}

\item{df2}{Analyseset of subset van de analyseset waar de doorstroom variabele consistent gemaakt dient te worden.
Dit dataframe bevat de variabelen INS_Studentnummer, INS_Opleidingsfase_BPM, INS_Datum_diploma, INS_Opleidingsnaam_2002,
INS_Studiejaar, INS_Eerste_jaar_opleiding_en_instelling en
INS_Doorstroom_van_bachelor_naar_master of INS_Doorstroom_van_premaster_naar_master.}

\item{bachelor_naar_master}{Standaard op TRUE, wanneer de variabele INS_Doorstroom_van_premaster_naar_master aangepast moet worden,
dan moet deze variabele op FALSE gezet worden.}
}
\value{
Aangepaste analyseset of subset van de analyseset.
}
\description{
Omdat de variabelen INS_Doorstroom_van_bachelor_naar_master en INS_Doorstroom_van_premaster_naar_master
enkele doorstromers niet juist categoriseren, maar consistentie van deze variabelen vereist is voor
het maken van een dashboard, wordt deze variabele aangepast op basis van een dataframe bepaald door
bepaal_doorstroom. Studenten die geen master volgen worden gecorrigeerd.
}
\seealso{
Other doorstroom: 
\code{\link{bepaal_doorstroom}()},
\code{\link{maak_doorstroomvar_premaster_master}()}
}
\concept{doorstroom}
