% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rstudio_prefs_silent.R
\name{pretty_print_updates}
\alias{pretty_print_updates}
\title{Pretty Print RStudio Preference Updates}
\usage{
pretty_print_updates(old, new)
}
\arguments{
\item{old}{A list of old RStudio preferences.}

\item{new}{A list of new RStudio preferences.}
}
\value{
A logical value indicating if there were any updates.
}
\description{
A function that creates a formatted output of RStudio preference updates, comparing old and new values.
}
\details{
This function generates a data frame with old and new preferences, pads each column with trailing spaces for alignment, and prints updates. It also returns a logical value indicating if there were any updates.
}
\note{
This function is useful for comparing and documenting changes in RStudio preferences.
}
