% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hash_var.R
\name{random_string_vector}
\alias{random_string_vector}
\title{Generate a random string vector}
\usage{
random_string_vector(n = 500, length, characters = c(letters[1:6], 0:9))
}
\arguments{
\item{n}{The number of items in the vector}

\item{length}{the number of characters in a string}

\item{characters}{A vector containing the characters to include}
}
\description{
Generate a random string vector
}
