% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_documentation.R
\name{read_data_dictionary}
\alias{read_data_dictionary}
\title{read_data_dictionary}
\usage{
read_data_dictionary(filename = "", metadata_location = NULL, ...)
}
\arguments{
\item{filename}{The filename of the data dictionary file.}

\item{metadata_location}{Location of the data dictionary files,
if NULL system variables will be used.}

\item{...}{Additional parameters to be passed to the file reading function.}
}
\value{
A data frame containing the data dictionary.
}
\description{
A function that reads a data dictionary file.
}
