% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_documentation.R
\name{read_documentation}
\alias{read_documentation}
\title{read Documentation}
\usage{
read_documentation(
  filename = "",
  documentation_location = NULL,
  readr = F,
  ...
)
}
\arguments{
\item{filename}{The filename}

\item{documentation_location}{Location of documentation files,
if NULL system variables will be used.}

\item{readr}{Defaults to FALSE, if TRUE then it uses
of 'read_delim'.}

\item{...}{The other parameters are passed to the function}
}
\value{
Imported data object
}
\description{
A generic function that can read a documentation file.
}
