% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_save.R
\name{read_file_proj}
\alias{read_file_proj}
\title{Read File Project}
\usage{
read_file_proj(
  name,
  settings_df = load_and_save_config_proj(),
  settings_type = NULL,
  dir = NULL,
  extension = NULL,
  fix_encoding = FALSE,
  encoding = "latin1",
  sub_dir = NA_character_,
  csv_readr_function_args = list(trim_ws = T, delim = ";", locale =
    readr::locale(decimal_mark = ",", grouping_mark = ".")),
  ...
)
}
\arguments{
\item{name}{Character, name of the file to be read.}

\item{settings_df}{Data frame containing settings, defaults to the result of \code{load_and_save_config_proj()}.}

\item{settings_type}{Character, the type of setting to use, default is NULL.}

\item{dir}{Character, directory path, default is NULL.}

\item{extension}{Character, file extension, default is NULL.}

\item{fix_encoding}{Logical, whether to fix encoding, default is FALSE.}

\item{encoding}{Character, encoding to be used, default is "latin1".}

\item{sub_dir}{Character, sub-directory, default is NA_character_.}

\item{csv_readr_function_args}{List, additional arguments for reading CSV files, default specified.}

\item{...}{Additional arguments to be passed to the reading functions.}
}
\value{
The data read from the specified file.
}
\description{
Reads a file from the project with extension-based reading functions, and applies specified encoding fixes if necessary.
Utilizes both manual and configuration settings for reading the file.
}
