% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_documentation.R
\name{read_import_definitions}
\alias{read_import_definitions}
\title{read_import_definitions}
\usage{
read_import_definitions(filename = "", metadata_location = NULL, ...)
}
\arguments{
\item{filename}{The filename of the import definitions file.}

\item{metadata_location}{Location of the import definitions files,
if NULL system variables will be used.}

\item{...}{Additional parameters to be passed to the file reading function.}
}
\value{
A data frame containing the import definitions.
}
\description{
A function that reads an import definitions file.
}
