% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{read_kolomwaarden}
\alias{read_kolomwaarden}
\title{Read kolomwaarden}
\usage{
read_kolomwaarden(Naam_bestand)
}
\arguments{
\item{Naam_bestand}{Naam van het bestand waar de kolomwaarden voor gebouwd worden,
zonder bestandssoort (dus enkel "Namen" of "Gegevens"), maar met aanhalingstekens}
}
\value{
Een tibble met een kolom met Kolomnaam en een kolom met Waarde
}
\description{
Functie die het kolomwaarden bestand van de opgegeven bestandsnaam uitleest en
doorgeeft als een tibble. Deze tibble bestaat uit Kolomnaam met bijbehorende
Waarden
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{test_naming}()},
\code{\link{veld_is_subset}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
