% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{read_kolomwaarden_testen}
\alias{read_kolomwaarden_testen}
\title{Read kolomwaarden testen}
\usage{
read_kolomwaarden_testen(Testdoc)
}
\arguments{
\item{Testdoc}{Testdocumentatie}
}
\description{
Inlezen van de kolomwaarden bij het testen met de testdocumentatie
}
