% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readrds_csv.R
\name{readrds_csv}
\alias{readrds_csv}
\title{Read RDS CSV}
\usage{
readrds_csv(
  dataloc = "",
  output,
  readr = FALSE,
  print = FALSE,
  fix.encoding = FALSE,
  encoding = "latin1",
  ...
)
}
\arguments{
\item{dataloc}{Data location on the network drive, starting from "G:/DSZ/SA2016/Datasets/"}

\item{output}{Data location on the network drive, depending on the branch
set.starting from "G:/DSZ/SA2016/Datasets/Output/(branch)/"}

\item{readr}{Use readr to read CSV files}

\item{print}{TRUE/FALSE: Print the date on which the imported file was last modified}

\item{fix.encoding}{TRUE/FALSE: Apply an encoder to the columns}

\item{encoding}{Which encoding to use}

\item{...}{Pass arguments for readr, read.csv2, fst or readRDS}
}
\value{
Dataframe
}
\description{
a generic function that can read an R object into an RDS, fst and CSV.
First determine the extension and then apply the appropriate read function
}
