% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tableau.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns in a dataframe}
\usage{
rename_columns(df, matching_cols, dfTableau_naming)
}
\arguments{
\item{df}{A dataframe.}

\item{matching_cols}{A character vector of column names in df that match with the mapping file.}

\item{dfTableau_naming}{A dataframe containing the mapping between old and new column names.}
}
\value{
The dataframe with renamed columns.
}
\description{
This function takes a dataframe and a list of matching columns, along with a mapping file,
and renames the columns in the dataframe based on the mapping file.
}
\examples{
\dontrun{
df <- rename_columns(df, matching_cols, dfTableau_naming)
}
}
