% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_metadata.R
\name{retrieve_data_metadata}
\alias{retrieve_data_metadata}
\title{Check classes}
\usage{
retrieve_data_metadata(filepath, verbose = FALSE)
}
\arguments{
\item{filepath}{Path to data file}

\item{verbose}{whether to print number of columns and number of rows.}
}
\value{
Column names
}
\description{
Retrieve number of rows and number of columns of a data file without reading it in.
Additionally, for csv files, it is possible to retrieve the column names.
Todo: make column names work for other file types as well.
}
