% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{return_assertions_message}
\alias{return_assertions_message}
\title{Retourneer de assertion message van een collectie}
\usage{
return_assertions_message(
  Collection,
  Name_collection,
  fail = "stop",
  silent = F,
  Outputmap = NULL
)
}
\arguments{
\item{Collection}{Een object met de classe AssertCollection}

\item{Name_collection}{De naam van de collectie. Deze naam wordt in de berichten
genoemd}

\item{fail}{"stop": Bij het falen van de assertions wordt een error gegeven en
de uitvoer van het script gestopt. Bij "warning" wordt alleen een warning gegeven}

\item{silent}{Bij FALSE (default) wordt de succesmelding in de console geprint. Bij TRUE wordt deze niet getoond}

\item{Outputmap}{Map, zoals 1. Ingelezen data, waarin bestand staat}
}
\value{
De melding of de assertion-test wel of niet geslaagd is
}
\description{
Geef een bericht over het wel of niet slagen van de assertion test. Hiervoor
moet een "assertion collection" van het package checkmate opgegeven worden.
Dit bericht wordt teruggegeven in de vorm van een error of een warning.
}
\details{
Voor enkele assertions zijn alleen warnings toegestaan, aangezien door een error
een script stopt met runnen. Dit is gedaan voor de assertions: percentage missende
waarden, dubbelingen, subset en set_equal.
}
