% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{return_assertions_message_testen}
\alias{return_assertions_message_testen}
\title{Return assertions message testen}
\usage{
return_assertions_message_testen(
  Collection,
  Name_collection,
  silent = F,
  Outputmap = NULL
)
}
\arguments{
\item{Collection}{Een object met de classe AssertCollection}

\item{Name_collection}{De naam van de collectie. Deze naam wordt in de berichten
genoemd}

\item{silent}{Bij FALSE (default) wordt de succesmelding in de console geprint. Bij TRUE wordt deze niet getoond}

\item{Outputmap}{Map, zoals 1. Ingelezen data, waarin bestand staat}
}
\value{
De melding of de assertion-test wel of niet geslaagd is
}
\description{
Geef een dataframe over het wel of niet slagen van de test. Hiervoor
moet een "assertion collection" van het package checkmate opgegeven worden
}
