% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saverds_csv.R
\name{sa_defineer_test_documentatie}
\alias{sa_defineer_test_documentatie}
\title{sa_defineer_test_documentatie}
\usage{
sa_defineer_test_documentatie(value = TRUE)
}
\arguments{
\item{value}{Logical: Whether to edit the test documentation.}
}
\description{
Global option to generate a new documentation file each time saverds_csv is used.
}
