% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saverds_csv.R
\name{sa_defineer_test_documentatie_get}
\alias{sa_defineer_test_documentatie_get}
\title{sa_defineer_test_documentatie}
\usage{
sa_defineer_test_documentatie_get()
}
\description{
Internal function read what was set in sa_defineer_test_documentatie()
}
