% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_voorbereidingen.R
\name{sa_voorbereidingen}
\alias{sa_voorbereidingen}
\title{SA Voorbereidingen.}
\usage{
sa_voorbereidingen(location = "99. Functies & Libraries/00. Voorbereidingen.R")
}
\arguments{
\item{location}{Location of the preparation script.}
}
\description{
Run script to set project requirements.
}
