% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fst.R
\name{save_fst}
\alias{save_fst}
\title{save fst}
\usage{
save_fst(
  Object_to_save,
  Name_to_save,
  output = "1. Ingelezen data/",
  dataloc,
  session_info = FALSE,
  ...
)
}
\arguments{
\item{Object_to_save}{Data frame.}

\item{Name_to_save}{File name to write without extension.}

\item{output}{Location on network drive, depending on the value of branch
"G:/DSZ/SA2016/Datasets/Output/(branch)/"}

\item{dataloc}{Location on network drive, having as root "G:/DSZ/SA2016/Datasets/"}

\item{session_info}{Default: FALSE. Whether to save session info to same directory.}

\item{...}{dot parameters to parse to write_fst}
}
\description{
write R data frame as fst..
}
