% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_multiple_tableau.R
\name{save_multiple_tableau}
\alias{save_multiple_tableau}
\title{Save to mulitple tableau drectories}
\usage{
save_multiple_tableau(
  Object_to_save,
  filepaths,
  overwrite = TRUE,
  save_fst = FALSE,
  save_csv = FALSE,
  fileEncoding = ""
)
}
\arguments{
\item{Object_to_save}{The object to save}

\item{filepaths}{To the datalocation of the project}

\item{overwrite}{whether to overwrite existing file when copying to other directory.}

\item{save_fst}{Default FALSE: No fst is written. When TRUE becomes}

\item{save_csv}{Default FALSE: No CSV is saved. When TRUE becomes
both a .rds and a .csv are saved.}

\item{fileEncoding}{The fileencoding for write.csv2 can be passed here.
If not specified, the default fileencoding will be used.}
}
\description{
A generic function to write data to multiple locations in different formats,
these being rds, fst and csv.
}
