% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savecsv.R
\name{savecsv}
\alias{savecsv}
\title{Is a generic function that stores an R object in CSV.}
\usage{
savecsv(
  Object_to_save,
  Name_to_save,
  output = "1. Ingelezen data/",
  dataloc,
  fileEncoding = "",
  session_info = FALSE
)
}
\arguments{
\item{Object_to_save}{The object to save}

\item{Name_to_save}{The file name that the file should be given (without extension)}

\item{output}{Data location on the network drive, depending on the branch
set.starting from "G:/DSZ/SA2016/Datasets/Output/(branch)/"}

\item{dataloc}{Data location on the network drive, starting from "G:/DSZ/SA2016/Datasets/"}

\item{fileEncoding}{If not specified, UTF-8 is used}

\item{session_info}{Default FALSE. Must the session info be in the same folder
be written off}
}
\description{
Is a generic function that stores an R object in CSV.
}
