% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savefst_csv.R
\name{savefst_csv}
\alias{savefst_csv}
\title{Save FST / CSV}
\usage{
savefst_csv(
  Object_to_save,
  Name_to_save,
  output = "1. Ingelezen data/",
  dataloc,
  fileEncoding = "",
  save_csv = FALSE,
  session_info = FALSE,
  Test_documentatie = FALSE,
  Documentatie = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{Object_to_save}{Het object dat opgeslagen moet worden}

\item{Name_to_save}{De bestandsnaam dat het bestand moet krijgen (zonder extensie)}

\item{output}{Datalocatie op de netwerkschijf, afhankelijk van de branch
die is ingesteld.startend van "G:/DSZ/SA2016/Datasets/Output/(branch)/"}

\item{dataloc}{Datalocatie op de netwerkschijf, startend van "G:/DSZ/SA2016/Datasets/"}

\item{fileEncoding}{De fileencoding voor write.csv2 kan hier meegegeven worden.
Als dit niet opgegeven wordt, wordt de standaard fileencoding gebruikt.}

\item{save_csv}{Standaard FALSE: Er wordt geen CSV weggeschreven. Als TRUE wordt
zowel een .rds als een .csv weggeschreven.}

\item{session_info}{Standaard FALSE. Moet de sessie-info in dezelfde folder
weggeschreve worden}

\item{Test_documentatie}{boolean die aangeeft of test documentatie gemaakt moet worden}

\item{Documentatie}{boolean die aangeeft of documentatie gemaakt moet worden}

\item{overwrite}{whether to overwrite existing file when copying to other directory.}
}
\description{
Function to write dataframes as fst and csv files on the shared network drive.
Default behaviour is to write file as fst. Optional is to write the file as csv as well.
}
