% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saverda.R
\name{saverda}
\alias{saverda}
\title{Save RDA}
\usage{
saverda(
  Object_to_save,
  Name_to_save,
  output = "1. Ingelezen data/",
  dataloc,
  session_info = FALSE
)
}
\arguments{
\item{Object_to_save}{The object to save}

\item{Name_to_save}{The file name that the file should be given (without extension)}

\item{output}{Data location on the network drive, depending on the branch
set.starting from "G:/DSZ/SA2016/Datasets/Output/(branch)/"}

\item{dataloc}{Data location on the network drive, starting from "G:/DSZ/SA2016/Datasets/"}

\item{session_info}{Default FALSE. Must the session info be in the same folder
be written off}
}
\description{
a function that saves an R object as an .rda file. Before the file
is stored, it is first transformed to a data.frame, so that
Tableau can read the file
}
