% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saverds_csv.R
\name{saverds_csv}
\alias{saverds_csv}
\title{Save RDS / CSV / FST}
\usage{
saverds_csv(
  Object_to_save,
  Name_to_save,
  output = "1. Ingelezen data/",
  dataloc,
  fileEncoding = "",
  save_csv = FALSE,
  save_fst = FALSE,
  session_info = FALSE,
  Test_documentatie = FALSE,
  Documentatie = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{Object_to_save}{Het object dat opgeslagen moet worden}

\item{Name_to_save}{De bestandsnaam dat het bestand moet krijgen (zonder extensie)}

\item{output}{Datalocatie op de netwerkschijf, afhankelijk van de branch
die is ingesteld.startend van "G:/DSZ/SA2016/Datasets/Output/(branch)/"}

\item{dataloc}{Datalocatie op de netwerkschijf, startend van "G:/DSZ/SA2016/Datasets/"}

\item{fileEncoding}{De fileencoding voor write.csv2 kan hier meegegeven worden.
Als dit niet opgegeven wordt, wordt de standaard fileencoding gebruikt.}

\item{save_csv}{Standaard FALSE: Er wordt geen CSV weggeschreven. Als TRUE wordt
zowel een .rds als een .csv weggeschreven.}

\item{save_fst}{Standaard FALSE: Er wordt geen fst weggeschreven. Als TRUE wordt
zowel een .rds als een .fst weggeschreven.}

\item{session_info}{Standaard FALSE. Moet de sessie-info in dezelfde folder
weggeschreve worden}

\item{Test_documentatie}{boolean die aangeeft of test documentatie gemaakt moet worden}

\item{Documentatie}{boolean die aangeeft of documentatie gemaakt moet worden}

\item{overwrite}{whether to overwrite existing file when copying to other directory.}
}
\description{
Een generieke functie die een R object in zowel een RDS, CSV als fst opslaat.
}
