% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_job.R
\name{schedule_job}
\alias{schedule_job}
\title{schedule job}
\usage{
schedule_job(file)
}
\arguments{
\item{file}{The relative path to the script in the project.}
}
\description{
Schedule a job on the R server by writing file paths to a txt file
on the shared network drive.
}
