% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sculpt_analysis_set.R
\name{sculpt_analysis_set}
\alias{sculpt_analysis_set}
\title{Sculpt Analysis set}
\usage{
sculpt_analysis_set(Analysis_set, file, Extra_sleutels = NULL, slack = T)
}
\arguments{
\item{Analysis_set}{De Analyseset}

\item{file}{De locatie van het script}

\item{Extra_sleutels}{Variabelenaam als textwaarde, of vector met meerdere
textwaaren van variabelenamen. Dn die gebruikt worden bij het uitvoeren van het
toevoegscript}

\item{slack}{Bij TRUE(standaard), worden debugging berichten ook naar Slack gestuurd.}
}
\description{
Wrapper rondom source() om de analyseset op te bouwen. Hierin wordt de analyseset
opgegeven en de naam van het script dat gesourced moet worden.
De functie maakt een kleine analyseset op basis van de volgende koppelvelden:
\itemize{
\item INS_Studentnummer
\item INS_Opleidingsnaam_2002
\item INS_Inschrijvingsjaar
\item INS_Studiejaar
\item INS_Inschrijvingsjaar_EOI
\item INS_Opleidingsfase_BPM
\item INS_ID_Student_Opleiding_Jaar
\item INS_ID_Student_Opleiding_Jaar_EOI
\item Optioneel: De velden opgegeven in Extra_sleutels
}
Daarnaast kunnen extra variabelen opgegeven worden als die gebruikt worden
in het script met behulp van de Extra_sleutels parameter.
}
