% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sculpt_as.R
\name{sculpt_as}
\alias{sculpt_as}
\title{sculpt_as}
\usage{
sculpt_as(Analysis_set, files_list)
}
\arguments{
\item{Analysis_set}{the Analysis set}

\item{files_list}{list of files to run}
}
\value{
Analysis set
}
\description{
Funtion to sculpt the analysis set.
}
