% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_delete}
\alias{sftp_delete}
\title{SFTP Delete}
\usage{
sftp_delete(remote_path, server, username, .opts)
}
\arguments{
\item{remote_path}{Het pad van het bestand op de server (exclusief server url)}

\item{server}{het adres van de server, zonder protocol ervoor, en zonder slash op het einde. (dus exclusief "sftp://")}

\item{username}{De gebruikersnaam die gebruikt wordt voor authenticatie op de server}

\item{.opts}{Een named lijst met opties die meegegeven worden aan de ftpUpload
functie. De opties die voor ssh authenticatie meegegeven moeten worden zijn
"ssh.public.keyfile" en "ssh.private.keyfile".}
}
\description{
Functie om bestanden te verwijderen van een remote sftp server. Dit is een helper
functie, die gebruikt wordt in sftp_vu_delete.
}
