% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_list_dir}
\alias{sftp_list_dir}
\title{SFTP list directory}
\usage{
sftp_list_dir(remote_path, server, username, .opts)
}
\arguments{
\item{remote_path}{Het pad van de folder op de server (exclusief server url)}

\item{server}{het adres van de server, zonder protocol ervoor, en zonder slash
op het einde. (dus exclusief "sftp://").}

\item{username}{De gebruikersnaam die gebruikt wordt voor authenticatie op de server.}

\item{.opts}{Een named lijst met opties die meegegeven worden aan de getBinaryURL
functie. De opties die voor ssh authenticatie meegegeven moeten worden zijn
"ssh.public.keyfile" en "ssh.private.keyfile".}
}
\value{
Een vector met de paden van alle bestanden in de opgegeven folder.
}
\description{
Een functie om de bestandsnamen van een map op een sftp server te verkrijgen.
Alle bestanden in een folder worden opgehaald en in een vector geretourneerd.
Dit is een helper functie die wordt gebruikt binnen sftp_vu_list_dir.
}
