% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_login_credentials_get}
\alias{sftp_login_credentials_get}
\title{Get SFTP Login credentials}
\usage{
sftp_login_credentials_get()
}
\description{
De instellingen voor de sftp-login credentials worden uit de opties
van de sessie gehaald. Deze opties kunnen worden gebruikt voor de sftp functies.
}
