% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_login_credentials_set}
\alias{sftp_login_credentials_set}
\title{Set SFTP Login credentials}
\usage{
sftp_login_credentials_set(
  username,
  server,
  ssh.public.keyfile,
  ssh.private.keyfile
)
}
\arguments{
\item{username}{De gebruikersnaam om in te loggen op de server}

\item{server}{Het adres van de server die benaderd moet worden (bijvoorbeeld: "server.domein.nl")}

\item{ssh.public.keyfile}{De locatie van het publieke ssh-keyfile}

\item{ssh.private.keyfile}{De locatie van het prive ssh-keyfile}
}
\description{
De instellingen voor de sftp-login credentials worden opgeslagen in de opties
van de sessie. Deze opties worden gebruikt voor de sftp functies.
}
