% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_vu_download}
\alias{sftp_vu_download}
\title{SFTP VU Download}
\usage{
sftp_vu_download(remote_path, local_path, return_bin = T)
}
\arguments{
\item{remote_path}{Het pad van het bestand op de server (exclusief server url "Studentanalytics.sftp.vu.nl")}

\item{local_path}{Optioneel kan een lokaal pad opgegeven worden waarnaar het
gedownloade bestand weggeschreven wordt. Als deze parameter leeggelaten wordt, wordt
dit bestand niet weggeschreven.}

\item{return_bin}{Wordt het gedownloadde bestand ook teruggegeven als binary? de
geretourneerde waarde is daarna in te lezen met inleesfuncties. (default = TRUE)}
}
\description{
Functie om bestanden te downloaden van een de VU sftp server. Het serveradres
en de gebruikersnaam van StudentAnalytics zijn hier al ingevoerd. De functie gaat
er van uit dat de public en private ssh key op deze locaties staan:
\itemize{
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa.pub
}
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa
}
}
}
