% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_vu_download_and_delete}
\alias{sftp_vu_download_and_delete}
\title{SFTP VU Download and delete}
\usage{
sftp_vu_download_and_delete(
  remote_path,
  local_folder,
  overwrite = F,
  delete = T
)
}
\arguments{
\item{remote_path}{Het pad van het bestand op de server dat gedownload moet worden
(exclusief server url "Studentanalytics.sftp.vu.nl")}

\item{local_folder}{De lokale folder waarin het bestand van de server weggeschreven
moet worden}

\item{overwrite}{FALSE is standaard. Als TRUE, dan wordt het lokale bestand overschreven
met het nieuwe bestand op de server.}

\item{delete}{TRUE is standaard. Het bestand wordt van de server verwijderd. Als FALSE wordt het
niet van de server verwijderd, maar wel gedownload.}
}
\description{
Functie om bestanden te downloaden en daarna te verwijderen van
de VU sftp server. Het serveradres en de gebruikersnaam van StudentAnalytics
zijn hier al ingevoerd. De functie gaat er van uit dat de public en private ssh
key op deze locaties staan:
\itemize{
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa.pub
}
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa
}
}
}
