% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_vu_list_dir}
\alias{sftp_vu_list_dir}
\title{SFTP VU list directory}
\usage{
sftp_vu_list_dir(remote_path)
}
\arguments{
\item{remote_path}{Het pad van het bestand op de server (exclusief server url
"Studentanalytics.sftp.vu.nl")}
}
\value{
Een vector met de paden van alle bestanden in de opgegeven folder.
}
\description{
Een functie om de bestandsnamen van een map op een sftp server te verkrijgen.
Alle bestanden in een folder worden opgehaald en in een vector geretourneerd.
Het serveradres en de gebruikersnaam van StudentAnalytics zijn hier al ingevoerd.
De functie gaat er van uit dat de public en private ssh key op deze locaties staan:
\itemize{
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa.pub
}
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa
}
}
}
