% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftp.R
\name{sftp_vu_upload}
\alias{sftp_vu_upload}
\title{SFTP VU Upload}
\usage{
sftp_vu_upload(local_path, remote_path)
}
\arguments{
\item{local_path}{Het lokale pad van het bestand dat geupload wordt}

\item{remote_path}{Het pad waarnaar het bestand op de server geupload moet worden
(exclusief server url "Studentanalytics.sftp.vu.nl")}
}
\description{
Functie om bestanden te uploaden naar een de VU sftp server. Het serveradres
en de gebruikersnaam van StudentAnalytics zijn hier al ingevoerd. De functie gaat
er van uit dat de public en private ssh key op deze locaties staan:
\itemize{
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa.pub
}
\item 
\enumerate{
\item Functies & Libraries/ssh/User keys/sa_rsa
}
}
}
