% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_vvconverter.R
\name{sort_columnnames}
\alias{sort_columnnames}
\title{Sort column names}
\usage{
sort_columnnames(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
New data frame with column names sorted.
}
\description{
Function to sort column names alphabetically.
}
