% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_columnnames.R
\name{sort_columnnames_analysis_set}
\alias{sort_columnnames_analysis_set}
\title{Sort column names analysis set}
\usage{
sort_columnnames_analysis_set(data)
}
\arguments{
\item{data}{The data frame.}
}
\value{
New data frame with column names sorted.
}
\description{
Function to sort analysis set column names by category and alphabetically.
The order of the categories is as follows:
\enumerate{
\item Keys to pair with
\item Registration fields
\item Pre-study data
\item Data from during the study
\item Alumni details
}
}
