% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_objects.R
\name{source_index}
\alias{source_index}
\title{Source script and clear its objects}
\usage{
source_index(..., script)
}
\arguments{
\item{...}{Objects to keep in the global environment which are created in the sourced scripts}

\item{script}{The script to source}
}
\description{
Source script and clear its objects
}
