% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{split_file}
\alias{split_file}
\title{Split a file into chunks based on empty lines}
\usage{
split_file(filename)
}
\arguments{
\item{filename}{A character string specifying the path to the file to be split.}
}
\value{
A list of character vectors, where each vector represents a chunk of the file.
}
\description{
This function reads a file, splits it into chunks based on empty lines,
and preserves the indentation of each line within a chunk.
}
