% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{subassertfunctie}
\alias{subassertfunctie}
\title{subassertfunctie}
\usage{
subassertfunctie(
  naming,
  assert_functie,
  df,
  collectie,
  selecteer_kolommen,
  veldtype,
  kolomwaarden = NULL
)
}
\arguments{
\item{naming}{documentatie met de te testen assertions}

\item{assert_functie}{meegegeven assertfunctie}

\item{df}{dataframe met de data die geassert wordt}

\item{collectie}{assertion-collectie}

\item{selecteer_kolommen}{De kolommen uit de naming die nodig zijn voor de assertion}

\item{veldtype}{Veldtype van de assertion in een character of een lijst met characters}

\item{kolomwaarden}{kolomwaarden om de assertions list, subset en set_equal mee
te controleren}
}
\description{
Functie die de meegegeven assertion checkt op basis van het meegegeven
namingbestand. Dit zijn o.a. numerieke, character en date assertions.
}
