% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_colnames_documentation.R
\name{tableau_colnames}
\alias{tableau_colnames}
\title{Translate Tableau colnames}
\usage{
tableau_colnames(
  dataframe,
  fieldname,
  fieldname_export,
  keep_other_colnames = T
)
}
\arguments{
\item{dataframe}{A dataframe for which column names need to be translated.}

\item{fieldname}{A column with fieldnames the dataframe will be changed to}

\item{fieldname_export}{a column with the original names of the dataframe, that wil be changed}

\item{keep_other_colnames}{Default TRUE, whether to keep empty columns names.}
}
\value{
a dataframe with translated column names
}
\description{
Fucntion to translate the columns of a dataset with regards to those in
the documentation file.
Empty column names are kept, for Tableau purposes
}
