% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{test_naming}
\alias{test_naming}
\title{Test naming}
\usage{
test_naming(df, Naming, Naam_bestand, Bericht = NULL, Outputmap = NULL)
}
\arguments{
\item{df}{Dataframe waar over geassert wordt}

\item{Naming}{Het bijbehorende documentatie bestand met assert informatie voor
het dataframe}

\item{Naam_bestand}{De naam van het databestand wat bekeken wordt. Deze naam moet
uniek zijn in het hele project}

\item{Bericht}{Bericht dat bij de assertion message bijgevoegd wordt.}

\item{Outputmap}{De Outputmap van het bestand, bijvoorbeeld 1. Ingelezen data}
}
\description{
Functie voor testen van een dataframe op basis van assertions.
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{read_kolomwaarden}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_pattern_table}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
