% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_colnames_documentation.R
\name{translate_colnames_documentation}
\alias{translate_colnames_documentation}
\title{Translate colnames documentation}
\usage{
translate_colnames_documentation(
  data,
  fieldname,
  fieldname_export,
  drop_na = T
)
}
\arguments{
\item{data}{A dataframe whose column names need to be translated}

\item{fieldname}{A column with fieldnames the dataframe will be changed to}

\item{fieldname_export}{a column with the original names of the dataframe, that wil be changed}

\item{drop_na}{Defaults to TRUE so that empty column names are dropped.}
}
\value{
a dataframe with translated column names
}
\description{
Function to translate the column names of a dataset to the columns contained in
the documentation file are defined. The empty column names will be removed
from the data.
}
