% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Up_to_date.R
\name{up_to_date}
\alias{up_to_date}
\title{up_to_date}
\usage{
up_to_date(bestandspad, correctie_tijdstip, frequentie, contact, inleesscript)
}
\arguments{
\item{bestandspad}{Bestandspad naar het ingelezen bestand}

\item{correctie_tijdstip}{Optioneel indien de data eerder geleverd kon worden.}

\item{frequentie}{De frequentie van de levering. Hier dient een integer opgegeven te worden die
de frequentie van de levering aangeeft. Bijvoorbeeld: Een jaarlijkse levering krijgt de waarde
365}

\item{contact}{Person to contact}

\item{inleesscript}{Naam van het inleesscript. Bijvoorbeeld "Inlezen Aanmeldingen.R"}
}
\description{
Check whether the read in data is up-to-date.
}
\details{
Function to be called inside of read scripts.
A slack message is sent when the data is not up-tp-date.
}
