% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rstudio_prefs_silent.R
\name{use_rstudio_prefs_silent}
\alias{use_rstudio_prefs_silent}
\title{Use RStudio Preferences Without Confirmation}
\usage{
use_rstudio_prefs_silent(...)
}
\arguments{
\item{...}{A series of RStudio preferences to update, e.g. \code{always_save_history = FALSE}, \code{rainbow_parentheses = TRUE}.}
}
\value{
NULL, updates RStudio \code{rstudio-prefs.json} file.
}
\description{
A wrapper for \code{rstudio.prefs::use_rstudio_prefs} that updates RStudio preferences without prompting for confirmation.
}
\details{
This function is designed to be used in startup scripts or automated processes where user confirmation is not desired. It bypasses the standard confirmation prompt by applying the preferences directly.
}
\note{
This function is a wrapper for \code{rstudio.prefs::use_rstudio_prefs} and is intended for use in environments where user interaction is not possible or desired.
}
\seealso{
\code{\link[rstudio.prefs]{use_rstudio_prefs}} for the original function that includes user confirmation.
}
