% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{validate_rm_objects}
\alias{validate_rm_objects}
\title{Valideer rm objecten}
\usage{
validate_rm_objects(filepath, export_naar_dataframe = FALSE)
}
\arguments{
\item{filepath}{Het complete filepath van het script}

\item{export_naar_dataframe}{Default: FALSE. Whether to export results to a data frame.}
}
\description{
Test die checkt of alle variabelen die worden aangemaakt ook weer
worden opgeruimd. De variabelen die niet worden opgeruimd, worden weergegeven
}
\seealso{
Other Script validatie: 
\code{\link{compare_input_output}()},
\code{\link{validate_assertions_present}()},
\code{\link{validate_clear_script_objects}()},
\code{\link{validate_introduction}()},
\code{\link{validate_no_warning_errors}()},
\code{\link{validate_script}()},
\code{\link{validate_style}()},
\code{\link{validate_write_files}()}
}
\concept{Script validatie}
