% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{validate_script}
\alias{validate_script}
\title{Valideer script}
\usage{
validate_script(filepath = NULL, export_naar_dataframe = FALSE)
}
\arguments{
\item{filepath}{De map en bestandsnaam van het script dat getest moet worden binnen het project.
Bijvoorbeeld: "01. Inlezen/Inlezen Afhakers.R". Als het filepath niet is
opgegeven, wordt het script dat in rstudio open staat gecontroleerd}

\item{export_naar_dataframe}{Logical value die aangeeft of de resultaten
van de functie in een dataframe geëxporteerd moeten worden}
}
\value{
Per test wordt het resultaat weergeven. Dit gebeurd in de volgorde zoals hierboven beschreven.
"Geslaagd" geeft aan dat een test zonder problemen is afgerond.
Mocht dit niet het geval zijn dan geeft een waarschuwing in het rood meer informatie.
De laatste output weergeeft eventuele scripts die afhankelijk zijn van het script in blauw.
}
\description{
Een functie voor het valideren van een script met behulp van een aantal tests.
}
\details{
Het script wordt gevalideerd op het volgende:
\itemize{
\item Geeft het script waarschuwingen of errors?
\item Zijn indien nodig assertions aanwezig?
\item Worden alle variabelen op het eind van het script opgeruimd?
\item Voldoen alle weggeschreven bestanden aan de documentatie?
\item Voldoet de introductie aan de stijlgids?
\item Volgt het script de stijlgids?
\item Welke scripts zijn afhankelijk van de output bestanden van het script?
}

Deze functie is primair geschikt voor de inlees en manipuleer scripts. Scripts
die afhankelijk zijn van eerder uitgevoerde scripts kunnen gedeeltelijk niet werken.

Let op: om de scripts te testen worden deze eenmaal uitgevoerd en worden bestanden
weg geschreven. Bij sommige scripts kan deze validatie daarom wat langer duren.
}
\seealso{
Other Script validatie: 
\code{\link{compare_input_output}()},
\code{\link{validate_assertions_present}()},
\code{\link{validate_clear_script_objects}()},
\code{\link{validate_introduction}()},
\code{\link{validate_rm_objects}()},
\code{\link{validate_write_files}()}
}
\concept{Script validatie}
