% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_proj.R
\name{validate_style}
\alias{validate_style}
\title{Validate script style}
\usage{
validate_style(filepath, export_to_dataframe = FALSE)
}
\arguments{
\item{filepath}{The complete filepath of the script.}

\item{export_to_dataframe}{A logical value indicating whether to export results to a data frame. Default is FALSE.}
}
\description{
This function tests a script for style errors using the lintr package.
}
\examples{
\dontrun{
# Validate the style of a script
validate_style(filepath = "/path/to/script.R")

# Validate the style of a script and export the results to a dataframe
validate_style(filepath = "/path/to/script.R", export_to_dataframe = TRUE)
}
}
\concept{Script validation}
