% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{variable_name_lint}
\alias{variable_name_lint}
\title{Lint function to check variable names}
\usage{
variable_name_lint()
}
\description{
Function to generate lint functionconcerning variable names not following the "stijlgids"
}
