% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentatie_tests.R
\name{veld_is_subset}
\alias{veld_is_subset}
\title{Veld is subset}
\usage{
veld_is_subset(Veld, df, Veldnamen, Veldtypes)
}
\arguments{
\item{Veld}{Naam van het te controleren veld}

\item{df}{Dataframe dat gecontroleerd word}

\item{Veldnamen}{Set waar het Veld een onderdeel van is}

\item{Veldtypes}{Types die bij de Veldnamen horen}
}
\description{
Functie die checkt of een veld bestaat uit subset-waarden. We defineren een subset
als het een characterveld is dat minder dan 20 unieke waarden bevat.
}
\seealso{
Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{test_naming}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}

Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{create_documentatie_tests}()},
\code{\link{find_patterns}()},
\code{\link{get_dist}()},
\code{\link{get_ratio}()},
\code{\link{match_kolom_met_patroon}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{vind_subset_kolommen}()},
\code{\link{write_documentatie_tests}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
\concept{tests}
