% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_assertions_message.R
\name{vind_assertion_soort}
\alias{vind_assertion_soort}
\title{Vind assertion soort}
\usage{
vind_assertion_soort(Regel)
}
\arguments{
\item{Regel}{Regel/character waarin gezocht wordt naar soort assertion}
}
\value{
Characters die in 1 woord omschrijven om welke assertion het gaat
}
\description{
Functie om per regel (assertion) te vinden welke specifieke assertion erbij hoort
}
