% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentatie_tests.R
\name{vind_kolomwaarden}
\alias{vind_kolomwaarden}
\title{Vind kolomwaarden}
\usage{
vind_kolomwaarden(Veldnaam, Veldtype, df)
}
\arguments{
\item{Veldnaam}{Veld/kolom waarvoor de kolomwaarden worden gezocht}

\item{Veldtype}{Type van het veld}

\item{df}{dataframe waarin de waarden worden gezocht}
}
\description{
Functie om kolomwaarden voor een veld te bepalen. Deze waarden zijn unieke
voorkomende waarden, muv NA.
}
