% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{vind_subset_kolommen}
\alias{vind_subset_kolommen}
\title{Vind subset kolommen}
\usage{
vind_subset_kolommen(Naming)
}
\arguments{
\item{Naming}{Een tabel of dataframe met de veldnamen van kolommen van het
desbetreffende dataframe met meer informatie.}
}
\value{
Veldnamen die type list hebben
}
\description{
Deze functie geeft de namen weer van de kolommen uit een dataframe die in gebruik
zijn en het veldtype subset (of list) hebben meegekregen. Dit wordt in character type gegeven.
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{write_pattern_table}()}
}
\concept{assertions}
