% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitiematrix.R
\name{voeg_transitie_subtotalen_toe}
\alias{voeg_transitie_subtotalen_toe}
\title{voeg_transitie_subtotalen_toe}
\usage{
voeg_transitie_subtotalen_toe(df_lang, df)
}
\arguments{
\item{df_lang}{Lange data frame (bvb output van bepaal_subtotalen_met_transitiematrix)}

\item{df}{Dataframe}
}
\value{
De transitiematrix na 'df \%>\% left_join(df_lang)' plus nog andere aanpassingen
}
\description{
Functie om de transitie aantallen toe te voegen aan de dataset (alleen nodig bij ongeneste data)
}
\examples{
\dontrun{voeg_transitie_subtotalen_toe(df_lang, df)}
}
