% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvconverter_data.R
\name{vvconverter_data}
\alias{vvconverter_data}
\title{Get path to vvconverter data}
\usage{
vvconverter_data(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
vvconverter comes bundled with a number of data files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
vvconverter_data()
vvconverter_data("Dates.csv")
}
