% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_outflow_study_year.R
\name{wrapper_determine_outflow_study_year}
\alias{wrapper_determine_outflow_study_year}
\title{determine_outflow_study_year}
\usage{
wrapper_determine_outflow_study_year(df)
}
\arguments{
\item{df}{Data frame}
}
\description{
Applies determine_outflow_study_year() for each row in a dataframe.
}
