% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_colnames_documentation.R
\name{wrapper_translate_colnames_documentation}
\alias{wrapper_translate_colnames_documentation}
\title{Wrapper translate colnames documentation}
\usage{
wrapper_translate_colnames_documentation(data, documentatie_df, drop_na = T)
}
\arguments{
\item{data}{A dataframe whose column names need to be translated}

\item{documentatie_df}{A dataframe with a fielname and fieldname_export column,
voor the translations of the colnames.}

\item{drop_na}{Defaults to TRUE so that empty column names are dropped.}
}
\value{
a dataframe with translated column names
}
\description{
Function that wraps the used variabalbes of SA-script around the translate
colnames documentation function.
}
