% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file.R
\name{write_file}
\alias{write_file}
\title{Save RDS / CSV / FST}
\usage{
write_file(
  Object_to_save,
  Name_to_save,
  destination = NULL,
  save_csv = FALSE,
  save_fst = FALSE,
  save_rds = FALSE,
  Documentatie = FALSE,
  show_rownames = FALSE
)
}
\arguments{
\item{Object_to_save}{Data object to write.}

\item{Name_to_save}{File name to use when writing (without extension).}

\item{destination}{Optional argument to set write location.}

\item{save_csv}{Default FALSE: whether to write file as a csv file.}

\item{save_fst}{Default FALSE: whether to write file as a fst file.}

\item{save_rds}{Default FALSE: whether to write file as a rds file.}

\item{Documentatie}{whether to create a documentation file.}

\item{show_rownames}{Whether to write row names in file}
}
\description{
Een generieke functie die een R object in zowel een RDS, CSV als fst opslaat.
}
