% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_save.R
\name{write_file_proj}
\alias{write_file_proj}
\title{Write File to Project}
\usage{
write_file_proj(
  object,
  name,
  settings_df = load_and_save_config_proj(),
  settings_type = NULL,
  dir = NULL,
  save_rds = NULL,
  save_csv = NULL,
  save_fst = NULL,
  sub_dir = NA_character_,
  rds_version = 2,
  csv_na = "",
  csv_sep = ";",
  csv_dec = ",",
  fst_compress = 100,
  ...
)
}
\arguments{
\item{object}{The object to be written.}

\item{name}{Character, name of the file.}

\item{settings_df}{Data frame containing settings, defaults to the result of \code{load_and_save_config_proj()}.}

\item{settings_type}{Character, the type of setting to use, default is NULL.}

\item{dir}{Character, directory path, default is NULL.}

\item{save_rds}{Logical, whether to save as RDS, default is NULL.}

\item{save_csv}{Logical, whether to save as CSV, default is NULL.}

\item{save_fst}{Logical, whether to save as FST, default is NULL.}

\item{sub_dir}{Character, sub-directory, default is NA_character_.}

\item{rds_version}{Numeric, version for RDS files, default is 2.}

\item{csv_na}{Character, NA representation for CSV files, default is "".}

\item{csv_sep}{Character, separator for CSV files, default is ";".}

\item{csv_dec}{Character, decimal separator for CSV files, default is ",".}

\item{fst_compress}{Numeric, compression level for FST files, default is 100.}

\item{...}{Additional arguments to be passed to the saving functions.}
}
\value{
NULL, the file is saved to the specified path.
}
\description{
Writes an object to a file in the project directory with specified saving options (RDS, CSV, or FST).
Utilizes both manual and configuration settings for writing the file.
}
