% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tableau.R
\name{write_tableau}
\alias{write_tableau}
\title{Write Tableau friendly dataframe}
\usage{
write_tableau(
  df,
  name,
  output_path,
  student_number_col = "INS_Studentnummer",
  enrolment_year_col = "INS_Inschrijvingsjaar_EOI",
  mapping_file = "Documentatie_Tableau_vriendelijke_variabelnamen_UT.csv",
  save_csv = TRUE,
  save_rds = FALSE,
  offset_years = 10
)
}
\arguments{
\item{df}{A dataframe.}

\item{name}{A character string specifying the name of the output files.}

\item{output_path}{A character string specifying the directory where the output files should be saved.}

\item{student_number_col}{Character string specifying the column name for student numbers. Default is "INS_Studentnummer".}

\item{enrolment_year_col}{Character string specifying the column name for enrolment years. Default is "INS_Inschrijvingsjaar_EOI".}

\item{mapping_file}{Character string specifying the name of the mapping file. Default is "Documentatie_Tableau_vriendelijke_variabelnamen_UT.csv".}

\item{save_csv}{Logical indicating whether to save the output as a CSV file. Default is TRUE.}

\item{save_rds}{Logical indicating whether to save the output as an RDS file. Default is FALSE.}

\item{offset_years}{An integer specifying the range of years to consider when filtering the 'INS_Inschrijvingsjaar_EOI' column. Default is 10.}
}
\value{
The modified dataframe.
}
\description{
This function takes a dataframe, renames its columns based on a mapping file,
applies some transformations, and saves the result in a specified location.
}
\examples{
\dontrun{
write_tableau(df, "output", "/path/to/directory", save_csv = TRUE, save_rds = FALSE, offset_years = 10)
}
}
