% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vvmover_create_documentatie.R
\name{bereken_na_waarden}
\alias{bereken_na_waarden}
\title{Bereken NA waarden}
\usage{
bereken_na_waarden(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{Extra argumenten}
}
\description{
Functie om per kolom te berekenen wat het percentage NAs is. Inclusief marge
van 10 procent.
}
