% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentatie_tests.R
\name{create_documentatie}
\alias{create_documentatie}
\title{Create documentation}
\usage{
create_documentatie(df, Naam, Limietwaarden_bestandspad = NA)
}
\arguments{
\item{df}{The dataframe}

\item{Naam}{Name of the documentation file to create}

\item{Limietwaarden_bestandspad}{}
}
\value{
Table
}
\description{
Create documentation
}
