% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_operators.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate \code{\%in\%} operator}
\usage{
... \%notin\% NA
}
\arguments{
\item{...}{Other arguments passed to the function.}
}
\value{
A logical value indicating if the element is not present in the vector.
}
\description{
This operator negates the \code{\%in\%} operator in R. It checks if an element is not present in a vector.
}
\examples{
1 \%notin\% seq(5)  # FALSE
6 \%notin\% seq(5)  # TRUE
}
