% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentatie_tests.R
\name{max_num}
\alias{max_num}
\title{Maximum numeric}
\usage{
max_num(x, ...)
}
\arguments{
\item{x}{Vector of which minimum is wanted}

\item{...}{extra arguments}
}
\description{
Function finds maximum of numeric vector.
Else it returns NA
}
