% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentatie_tests.R
\name{maxnum}
\alias{maxnum}
\title{Maximum numeric}
\usage{
maxnum(x)
}
\arguments{
\item{x}{Vector of which minimum is wanted}
}
\description{
Function finds maximum of numeric vector.
Else it returns NA
}
