% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_vvmover_create_documentatie.R
\name{min_num}
\alias{min_num}
\title{Minimum numeric}
\usage{
min_num(x, ...)
}
\arguments{
\item{x}{Vector of which minimum is wanted}

\item{...}{extra arguments}
}
\description{
Function finds minimum of numeric vector. Else it returns NA
}
