% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentatie_tests.R
\name{minnum}
\alias{minnum}
\title{Minimum numeric}
\usage{
minnum(x)
}
\arguments{
\item{x}{Vector of which minimum is wanted}
}
\description{
Function finds minimum of numeric vector. Else it returns NA
}
