% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documentatie_tests.R
\name{round_values}
\alias{round_values}
\title{Round context
negative values rounded down, positive up}
\usage{
round_values(x)
}
\arguments{
\item{x}{numeric}
}
\value{
rounded value
}
\description{
Round context
negative values rounded down, positive up
}
