% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_validation.R
\name{validate_no_warning_errors}
\alias{validate_no_warning_errors}
\title{Valideer geen warnings errors}
\usage{
validate_no_warning_errors(filepath, export_naar_dataframe = FALSE)
}
\arguments{
\item{filepath}{Het complete filepath van het script dat getest moet worden}

\item{export_naar_dataframe}{Default: FALSE. Whether to export results to a data frame.}
}
\description{
Een test die checkt of een script waarschuwingen en/of errors geeft
}
\seealso{
Other Script validatie: 
\code{\link{compare_input_output}()},
\code{\link{validate_assertions_present}()},
\code{\link{validate_clear_script_objects}()},
\code{\link{validate_introduction}()},
\code{\link{validate_rm_objects}()},
\code{\link{validate_script}()},
\code{\link{validate_style}()},
\code{\link{validate_write_files}()}
}
\concept{Script validatie}
