% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assert_naming.R
\name{write_pattern_table}
\alias{write_pattern_table}
\title{Write patterns}
\usage{
write_pattern_table(pattern)
}
\arguments{
\item{pattern}{Patroon dat moet worden weggeschreven}
}
\description{
Functie die pattern wegschrijft naar csv bestand als deze er nog niet instaat
}
\seealso{
Other assertions: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{build_kolomwaarden}()},
\code{\link{check_assertions}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{read_kolomwaarden}()},
\code{\link{test_naming}()},
\code{\link{vind_subset_kolommen}()}

Other tests: 
\code{\link{assert_character_named}()},
\code{\link{assert_character_set_equal}()},
\code{\link{assert_character_subset}()},
\code{\link{assert_date_named}()},
\code{\link{assert_logical_named}()},
\code{\link{assert_naming}()},
\code{\link{assert_no_duplicates_in_group}()},
\code{\link{assert_numeric_named}()},
\code{\link{assert_posixct_named}()},
\code{\link{check_assertions}()},
\code{\link{check_education_name}()},
\code{\link{controleer_na_waarden}()},
\code{\link{find_patterns}()},
\code{\link{test_naming}()}
}
\concept{assertions}
\concept{tests}
